// src/models/IoK8sApiCoreV1ReplicationControllerList/types.ts
import { IoK8sApiCoreV1ReplicationControllerModelOptions as IoK8sApiCoreV1ReplicationController } from '../IoK8sApiCoreV1ReplicationController';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiCoreV1ReplicationControllerListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiCoreV1ReplicationController>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}