// src/models/IoK8sApiCoreV1ReplicationControllerList/index.ts
import { IoK8sApiCoreV1ReplicationControllerModelOptions as IoK8sApiCoreV1ReplicationController } from '../IoK8sApiCoreV1ReplicationController';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';
import { BaseModel } from '../types';
import type { IoK8sApiCoreV1ReplicationControllerListModelOptions } from './types';
import { validate } from './validation';

export * from './types';

export class IoK8sApiCoreV1ReplicationControllerListModel extends BaseModel {
  // APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
  apiVersion?: string;
  // List of replication controllers. More info: https://kubernetes.io/docs/concepts/workloads/controllers/replicationcontroller
  items: Array<IoK8sApiCoreV1ReplicationController>;
  // Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;

  constructor(data: IoK8sApiCoreV1ReplicationControllerListModelOptions) {
    super();
    validate(data);
    this.apiVersion = data.apiVersion;
    this.items = data.items;
    this.kind = data.kind;
    this.metadata = data.metadata;
  }

  static create(
    data: IoK8sApiCoreV1ReplicationControllerListModelOptions,
  ): IoK8sApiCoreV1ReplicationControllerListModel {
    return new IoK8sApiCoreV1ReplicationControllerListModel(data);
  }

  toObject(): Partial<this> {
    return {
      apiVersion: this.apiVersion,
      items: this.items,
      kind: this.kind,
      metadata: this.metadata,
    } as Partial<this>;
  }
}

export default IoK8sApiCoreV1ReplicationControllerListModel;
